+++ CHANGELOG.md
@@ -3,3 +3,5 @@
 - Command parsing
 - Command interpretation
 - Command execution
+- Show help command
+- Empty input handler